+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-### 0.0.0
+### 0.0.0 / 2025-05-02
 
 - Binary size/performance optimization of a very basic Ratatui app
 
 // 3.4M
 
-### 0.0.1
+### 0.0.1 / 2026-04-04
 
 - Add lints rules
 - Add basic `main`/`app` scaffolding and main loop
@@ -13,3 +13,12 @@
 - Add `anyhow` crate
 
 // 602K (615.592)
+
+### 0.0.2 / 2026-04-04
+
+- Introduce the Main Buffer `stderr` messages and improve error handling
+- Put log file in a more appropriate location: ~/.local/state/env!("CARGO_PKG_NAME")
+- Check a minimum number of terminal columns and rows
+- Refactor `main` and `logger`
+
+// 611K (625.328)